# openapi.yaml
openapi: 3.0.0
info:
  title: KST Hackathon App
  description: The KST Hackathon App.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /user-data/{userId}:
    get:
      summary: Get user data
      tags:
        - Users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User data found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserData'
        '404':
          description: User data not found

    put:
      summary: Update user data
      tags:
        - Users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDataUpdate'
      responses:
        '200':
          description: User data updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserData'
        '404':
          description: User data not found

  # Technology Stack
...
  /hackathons:
    get:
      summary: Get all hackathons
      tags:
        - Hackathons
      responses:
        '200':
          description: A list of hackathons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hackathon'

    post:
      summary: Create a new hackathon
      tags:
        - Hackathons
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HackathonCreate'
      responses:
        '201':
          description: Hackathon created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hackathon'
...
  /teams/{teamId}/members:
    get:
      summary: Get team members
      tags:
        - Teams
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Team members retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMember'

    post:
      summary: Add a member to team
      tags:
        - Teams
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberCreate'
      responses:
        '201':
          description: Team member added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'

  # Prize endpoints
...
  /hackathons/{hackathonId}/products/{productId}/evaluations:
    get:
      summary: Get evaluations for a product in a hackathon
      tags:
        - Evaluation
        - Products
        - Hackathons
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evaluation'

    post:
      summary: Create an evaluation for a product in a hackathon
      tags:
        - Evaluation
        - Products
        - Hackathons
      parameters:
        - name: hackathonId
          in: path
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationCreate'
      responses:
        '201':
          description: Evaluation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evaluation'
components:
  schemas:
...
    UserData:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: string
        affiliation:
          type: string
        email:
          type: string
          format: email
        profile:
          type: string
      required:
        - userId
        - userName
        - affiliation
        - email
        - profile
    UserDataUpdate:
      type: object
      properties:
        userName:
          type: string
        affiliation:
          type: string
        email:
          type: string
          format: email
        profile:
          type: string

    # Authentication related schemas
...
    Hackathon:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        theme:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - theme
        - startDate
        - endDate
        - createdAt
        - updatedAt
    HackathonCreate:
      type: object
      properties:
        name:
          type: string
        theme:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      required:
        - name
        - theme
        - startDate
        - endDate
...
    TeamMember:
      type: object
      properties:
        userId:
          type: string
        hackathonId:
          type: integer
        teamId:
          type: integer
      required:
        - userId
        - hackathonId
        - teamId
    TeamMemberCreate:
      type: object
      properties:
        userId:
          type: string
        hackathonId:
          type: integer
      required:
        - userId
        - hackathonId

    # Prize related schema
...
    Evaluation:
      type: object
      properties:
        userId:
          type: string
        hackathonId:
          type: integer
        productId:
          type: integer
      required:
        - userId
        - hackathonId
        - productId
    EvaluationCreate:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []